Remove header and add gt.
For bookdown, we don't need (nor want) separate yaml headers on each chapter.

Also, whenever you use a new package, type usethis::use_package("thepackage") in your console (for example, usethis::use_package("gt") in this case). That will add the package to the DESCRIPTION file so the book-builder knows to install it.